+++ packages/yoga/src/Card/native/GymCard/CheckIn/index.js
@@ -1,0 +1,3 @@
+import CheckIn from './CheckIn';
+
+export default CheckIn;