+++ packages/yoga/src/index.js
@@ -4,11 +4,12 @@ import List from './List';
 import Checkbox from './Checkbox';
 import Slider from './Slider';
 import RadioGroup from './RadioGroup';
-import { Card, PlanCard } from './Card';
+import { Card, PlanCard, EventCard } from './Card';
 import Stepper from './Stepper';
 import { Container, Row, Col, Hide } from './Grid';
 import Text from './Text';
 import Rating from './Rating';
+import Tag from './Tag';
 
 export {
   ThemeProvider,
@@ -27,4 +28,6 @@ export {
   Hide,
   Text,
   Rating,
+  EventCard,
+  Tag,
 };